feat: Worked on the setup command functions
Progress has been steady. I have an overall idea of what I want to do. This evening will be spent finishing the implementation and testing it out as it is now to see if it prints out errors correctly and does not break down. I believe I should be done by maybe Thursday with making all the functions needed to execute commands. Next week will dedicate on the small things and fixing up this weeks work. Almost done.